pyproject.toml
[project]
name = "letta-myADE"
version = "0.1.0"
description = "Letta Open ADE with a first-party OpenAI-compatible model router."
requires-python = ">=3.12,<3.13"
dependencies = [
    "fastapi>=0.135.2",
    "httpx>=0.28.1",
    "letta-client>=1.8.0",
    "pydantic-settings>=2.11.0",
    "python-dotenv>=1.2.2",
    "tenacity>=9.1.2",
    "uvicorn>=0.42.0",
]

[tool.uv]
package = false

[dependency-groups]
dev = [
    "ipykernel>=7.2.0",
    "ipywidgets>=8.1.8",
    "marimo[recommended]>=0.21.1",
    "pytest>=9.0.3",
]
